feat(v4.4.0): neural diffusion orchestrator wiring — phase_diffusion_enhancement honours config.diffusion.backend
Before v4.4.0, Phase 15 of the generation pipeline was hardcoded to
instantiate `StatisticalDiffusionBackend` regardless of what
`config.diffusion.backend` said. The `NeuralDiffusionBackend` had
existed since v4.2 with full training + sampling + save/load support,
but callers had to instantiate it themselves from a separate driver —
the orchestrator phase never loaded it.

v4.4.0 wires the config through end-to-end.

Orchestrator dispatch
---------------------

`phase_diffusion_enhancement` now reads `config.diffusion.backend`
and dispatches:

- `"statistical"` (default) → existing moment-matching path, unchanged
- `"neural"` / `"hybrid"` + `feature = "neural"` → new
  `run_neural_diffusion_phase(entries)`:
  - Extracts (total_amount, line_count, approval_level) rows from the
    first 5,000 JEs
  - Loads a checkpoint from
    `config.diffusion.neural.checkpoint_path` when set, else trains
    a fresh score network via `NeuralDiffusionTrainer::train`
  - Generates `config.diffusion.sample_size` samples and records the
    count in `EnhancedGenerationStatistics`
- `"neural"` / `"hybrid"` without the `neural` feature → warn +
  graceful fallback to statistical (so a config copied between build
  variants degrades rather than failing)
- anything else → warn + fallback to statistical

Config schema
-------------

Pre-existing `DiffusionSchemaConfig.backend: String` stays as-is
(already parsed to "statistical"/"neural"/"hybrid" pre-v4.4.0).
New addition: `NeuralDiffusionSchemaConfig.checkpoint_path:
Option<String>` for loading pre-trained models.

    diffusion:
      enabled: true
      backend: neural
      n_steps: 100
      sample_size: 500
      neural:
        hidden_dims: [256, 256, 128]
        training_epochs: 100
        checkpoint_path: /path/to/model  # optional

Cargo features
--------------

Added `neural` + `neural-cuda` passthrough features on both
`datasynth-runtime` and `datasynth-cli` so downstream consumers can
rebuild with `--features neural` / `--features neural-cuda` and get
the full stack enabled.

Verification
------------

- Default build: `cargo check --release -p datasynth-runtime` clean,
  neural path behind cfg gate, `entries` param marked unused via
  `#[cfg_attr(not(feature = "neural"), allow(unused_variables))]`.
- Neural build: `cargo check --release -p datasynth-runtime
  --features neural` clean.
- `cargo clippy --release -p datasynth-runtime -p datasynth-cli
  -- -D warnings` clean under default features.
- `cargo fmt --check` clean.
- New integration test
  `crates/datasynth-runtime/tests/neural_diffusion_wiring.rs`:
  - `statistical_backend_is_the_default` — drives the full
    orchestrator with backend=statistical, confirms samples
    produced. Marked `#[ignore]` (takes ~4 min end-to-end via
    demo_preset).
  - `neural_backend_trains_and_produces_samples` — feature-gated,
    marked `#[ignore]`, runs only when neural is compiled in.
  - `neural_backend_falls_back_when_feature_disabled` — compiles only
    when neural is OFF; verifies the fallback path.
  Invoke with `cargo test -- --ignored` when you want end-to-end
  coverage.
- Fixed a drive-by useless-conversion clippy lint in the SAF-T
  CLI path introduced in v4.3.1.

Documentation
-------------

`docs/book/src/ai/neural-diffusion.md` gains a "v4.4.0 — Orchestrator
wiring" section documenting the dispatch matrix, checkpoint usage,
and known follow-ups (neural samples still standalone, not fed back
into downstream generators; feature extraction limited to 3 features;
true weighted hybrid blending not yet implemented).

+++ crates/datasynth-cli/src/main.rs
@@ -1916,7 +1916,7 @@ fn main() -> Result<()> {
                         // year/month arithmetic; clamp day to 28 to avoid
                         // month-overrun on Feb.
                         let mut ey = sy as i32;
-                        let mut em = sm + u32::from(config_for_manifest.global.period_months);
+                        let mut em = sm + config_for_manifest.global.period_months;
                         while em > 12 {
                             ey += 1;
                             em -= 12;
